Move info card to design-system and a util to common package
