+++ packages/design-system/src/components/index.ts
@@ -30,6 +30,8 @@ export { default as CookieDetails } from './cookieDetails';
 export { default as Details } from './cookieDetails/details';
 export { default as CookieTable } from './cookieTable';
 export { default as LandingPage } from './landingPage';
+export { default as InfoCard } from './landingPage/infoCard';
+export * from './landingPage/infoCard/fetchPSInfo';
 export { default as ErrorFallback } from './errorFallback';
 export { default as ExtensionReloadNotification } from './errorFallback/extensionReloadNotification';
 export * from './table';
